makefile new

+++ src/SelectionPolicy.cpp
@@ -61,7 +61,7 @@ const FacilityType& BalancedSelection::selectFacility(const vector<FacilityType>
     int min_diff = std::max({cur_lq, cur_eco, cur_env}) - std::min({cur_lq, cur_eco, cur_env});
     int selected_fac_index = 0;
 
-    for(int i = 1; i < facilitiesOptions.size(); i++){
+    for(size_t i = 1; i < facilitiesOptions.size(); i++){
         cur_lq = std::abs(facilitiesOptions[i].getLifeQualityScore() + LifeQualityScore);
         cur_eco = std::abs(facilitiesOptions[i].getEconomyScore() + EconomyScore);
         cur_env = std::abs(facilitiesOptions[i].getEnvironmentScore() + EnvironmentScore);
